backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# OpenAI
openai==1.3.7

# Environment
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2

# Logging
structlog==23.2.0

# CORS
python-cors==1.0.0
